Added a beginning of an asset tool

+++ Lamp/src/Lamp/Rendering/Renderer.h
@@ -19,7 +19,7 @@ namespace Lamp
 		static void Begin(OrthographicCamera& camera);
 		static void End();
 
-		static void AddSprite(Sprite& sprite) { m_RenderSprites.push_back(sprite); sprite.SetPosition(m_RenderSprites.size() - 1); }
+		static void AddSprite(Sprite& sprite) { m_RenderSprites.push_back(sprite); sprite.SetPosition((uint32_t)m_RenderSprites.size() - 1); }
 		static void RemoveSprite(Sprite& sprite) { m_RenderSprites.erase(m_RenderSprites.begin() + sprite.GetPosition()); }
 
 		static void Draw(const std::shared_ptr<Shader>& shader, const Sprite& sprite);
